don't truncate atmospheres at 10000K, extrapolate molecular equlibrium constants, update tests
