+++ test/utilities.jl
@@ -12,8 +12,8 @@ Additionally, in cases where actual and reference don't have equal values but do
 shape, the locations of the max errors are also printed.
 
 # Keywords
-- `rtol`: The desired relative tolerance.  Use 0 to skip relative comparison.
-- `atol`: The desired absolute tolerance
+- `rtol`: The desired relative tolerance.  Use 0 to skip.
+- `atol`: The desired absolute tolerance.  Skipped by default.
 - `err_msg`: An optional error message that can be printed when the arrays are not close
 - `error_location_fmt`: An optional keyword used to specify a function that creates a string 
   describing the location in the arrays where values are unequal, given a `CartesianIndex` object.
